+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Unreleased
+* feat: add Asset Allocation tab to compare current vs. target allocation and suggest rebalancing
+
 ## v2.7.0 (2026-07-04)
 * feat: support multiple currencies per account [#190](https://github.com/andreasgerstmayr/fava-portfolio-returns/pull/190) ([@andreasgerstmayr](https://github.com/andreasgerstmayr))
 * Fixed investments filter [#188](https://github.com/andreasgerstmayr/fava-portfolio-returns/pull/188) ([@romamo](https://github.com/romamo))